Fixing up intree probing

+++ src/solver.h
@@ -232,7 +232,7 @@ class Solver : public Searcher
             const Lit lit1
             , const Lit lit2
             , const bool red
-            , const uint64_t ID
+            , const int32_t ID
             , [[maybe_unused]] const bool checkUnassignedFirst = true
         );
         void detach_bin_clause(
